feat: store humans in an actual database
This uses `sqlx` to read/write humans from/to the sample-db.

The schema consists of a custom enum type for episodes, and a table for
humans. For now, the same structs as were used for the GraphQL schema
are being used to (de)serialize database values. This would have been
quite smooth, but sadly, `sqlx`'s derivation can't handle `Vec`s of
custom enums (https://github.com/launchbadge/sqlx/issues/298), so we
have to jump through some hoops by introducing `EpisodeSlice` and
`EpisodeVec` newtypes, which can then implement the required traits
(plus all the required juniper traits, which is the bigger pain).

Since we're using `sqlx`'s macros to check queries at compile time, we
need to connect to the database during compilation. The macro will use
the `DATABASE_URL` environment variable, which it can read from a `.env`
file, so we now write one of these files as part of `make prepare-db`
(note that this is for the benefit of editors, language servers, etc.,
`make run` would already inherit the `DATABASE_URL` when compiling).